Add target selection
Instead of a mere keyword, an alias can be represented as a map of
`symbol` -> `keyword`.

First CLI argument can be a symbol which will effectively act as a
selector. This is useful for managing internal dependencies. Sometimes,
an internal dependency is needed as local source, sometimes as a
released jar.